Renamed Abstract Repository
